+++ R/node_binomial.r
@@ -2,8 +2,7 @@
 ## a node modeled using logistic regression
 #' @export
 node_binomial <- function(data, parents, formula=NULL, betas, intercept,
-                          return_prob=FALSE, coerce2factor=FALSE,
-                          coerce2numeric=FALSE, labels=NULL) {
+                          return_prob=FALSE, output="logical", labels=NULL) {
 
   if (!data.table::is.data.table(data)) {
     data.table::setDT(data)
@@ -21,18 +20,16 @@ node_binomial <- function(data, parents, formula=NULL, betas, intercept,
   prob <- 1/(1 + exp(-prob))
   out <- rbernoulli(n=nrow(data), p=prob)
 
-  if (!is.null(labels)) {
+  if (output %in% c("character", "factor") && !is.null(labels)) {
     out[out] <- labels[1]
     out[out!=labels[1]] <- labels[2]
   }
 
-  if (coerce2factor & is.null(labels)) {
+  if (output=="factor" & is.null(labels)) {
     out <- factor(out)
-  } else if (coerce2factor) {
+  } else if (output=="factor") {
     out <- factor(out, levels=labels)
-  }
-
-  if (is.null(labels) & !coerce2factor & coerce2numeric) {
+  } else if (output=="numeric") {
     out <- as.numeric(out)
   }