+++ tests/testthat/test_get_interaction_term_for_formula.r
@@ -6,7 +6,7 @@ test_that("general test cases", {
   dag <- empty_dag() +
     node("A", "rbernoulli") +
     node("B", "rnorm") +
-    node("C", "rcategorical", probs=c(0.2, 0.2, 0.8), coerce2factor=TRUE)
+    node("C", "rcategorical", probs=c(0.2, 0.2, 0.8), output="factor")
   data <- sim_from_dag(dag, n_sim=200)
 
   d_combs <- get_cat_col_levs(data)