+++ tests/testthat/test_node_binomial.r
@@ -40,42 +40,42 @@ test_that("getting probabilities", {
   expect_equal(stats::sd(dat$B), 0.02805198, tolerance=0.0001)
 })
 
-test_that("coerce2factor", {
+test_that("as factor", {
 
   set.seed(65346)
 
   dag <- empty_dag() +
     node("A", type="rnorm", mean=10, sd=3) +
     node("B", type="binomial", parents="A", betas=1, intercept=-6,
-         coerce2factor=TRUE)
+         output="factor")
 
   dat <- sim_from_dag(dag=dag, n_sim=100)
 
   expect_equal(levels(dat$B), c("FALSE", "TRUE"))
 })
 
-test_that("coerce2factor + labels", {
+test_that("as factor + labels", {
 
   set.seed(65346)
 
   dag <- empty_dag() +
     node("A", type="rnorm", mean=10, sd=3) +
     node("B", type="binomial", parents="A", betas=1, intercept=-6,
-         coerce2factor=TRUE, labels=c("male", "female"))
+         output="factor", labels=c("male", "female"))
 
   dat <- sim_from_dag(dag=dag, n_sim=100)
 
   expect_equal(levels(dat$B), c("male", "female"))
 })
 
-test_that("coerce2numeric", {
+test_that("as numeric", {
 
   set.seed(65456546)
 
   dag <- empty_dag() +
     node("A", type="rnorm", mean=10, sd=3) +
     node("B", type="binomial", parents="A", betas=1, intercept=-6,
-         coerce2numeric=TRUE)
+         output="numeric")
 
   dat <- sim_from_dag(dag=dag, n_sim=100)