+++ NEWS.md
@@ -58,6 +58,7 @@ Enhancements
 
 * Added the `output` argument to the `rbernoulli()` function to allow different output formats.
 * Change default of `sort_dag` in `sim_from_dag()` from `TRUE` to `FALSE`.
+* Moved the functionality of the `coerce2factor` and `coerce2numeric` arguments in `rcategorical()`, `node_multinomial()` and `node_binomial()` to the `output` argument for a more consistent syntax and easier usage.
 
 Bug Fixes